Discovered why l=32 was not working, we used uint32_t./long_message_attack_parallel  This commit has the base for benchmarking the filling rate for the dictionary

+++ long-message-attack/dict.h
@@ -46,7 +46,7 @@ typedef struct {
  
 
 typedef struct {
-  UINT nslots; // number of elements  in the dictionary
+  size_t nslots; // number of elements  in the dictionary
   slot* slots; // == slot slots[nslots] but on the heap
 } dict;
